+++ src/hypervisor.ts
@@ -1,16 +1,16 @@
 import {Architecture} from './architecture'
-import {Vm as QemuVm} from './qemu_vm'
 import {getOrDefaultOrThrow} from './utility'
 
 export enum Kind {
-  qemu
+  qemu,
+  simh
 }
 
 export interface Hypervisor {
   get kind(): Kind
   get sshPort(): number
   get firmwareFile(): string
-  get vmModule(): typeof QemuVm
+  get binaryDirectory(): string
   get efi(): Hypervisor
   getResourceUrl(architecture: Architecture): string
   resolve<T>(implementation: Record<string, T>): T
@@ -31,8 +31,8 @@ export class Qemu implements Hypervisor {
     return `${this.firmwareDirectory}/bios-256k.bin`
   }
 
-  get vmModule(): typeof QemuVm {
-    return QemuVm
+  get binaryDirectory(): string {
+    return 'bin'
   }
 
   get efi(): Hypervisor {
@@ -64,5 +64,37 @@ export class QemuRiscv extends Qemu {
 
   override get efi(): Hypervisor {
     return this
+  }
+}
+
+export class Simh implements Hypervisor {
+  get kind(): Kind {
+    return Kind.simh
+  }
+
+  get sshPort(): number {
+    return 2847
+  }
+
+  // SIMH simulators have their firmware built in.
+  get firmwareFile(): string {
+    return ''
+  }
+
+  // The simulator binary is located at the root of the archive.
+  get binaryDirectory(): string {
+    return ''
+  }
+
+  get efi(): Hypervisor {
+    throw Error('SIMH does not support EFI')
+  }
+
+  getResourceUrl(architecture: Architecture): string {
+    return architecture.resourceUrl
+  }
+
+  resolve<T>(implementation: Record<string, T>): T {
+    return getOrDefaultOrThrow(implementation, 'simh')
   }
 }