+++ src/operating_systems/netbsd/factory.ts
@@ -1,12 +1,19 @@
 import {OperatingSystem} from '../../operating_system'
+import {Class} from '../../utility'
 import {factory} from '../factory'
 import QemuFactory from '../qemu_factory'
 import NetBsd from './netbsd'
+import NetBsdVax from './vax'
 
 factory(
   class NetBsdFactory extends QemuFactory {
     override createImpl(version: string): OperatingSystem {
-      return new NetBsd(this.architecture, version)
+      const cls = this.architecture.resolve<Class<NetBsd>>({
+        vax: NetBsdVax,
+        default: NetBsd
+      })
+
+      return new cls(this.architecture, version)
     }
   }
 )